Wait for firefox nightly to appear
The au3 which installs firefox is run just before this one, and should
install firefox nightly.

However, since that one requires admin rights, it appears to exit
immediately (as far as buildbot is concerned). Therefore, installing the
addon will most likely fail, because we can't actually start firefox.

Fix by having the script poll for the existence of that file before
actually trying to do anything.