Add an extra delay
There seems to be some race between the closing of firefox nightly at
the end of the previous script, and the re-opening of it here.

Make sure that doesn't happen by ensuring there's at least a second
between the disappearance of the downloads window in the previous script
and the opening of firefox here.